+++ lib/aspera/cli/plugins/config.rb
@@ -718,7 +718,11 @@ module Aspera
               fields: Transfer::SpecDoc::TABLE_COLUMNS.map(&:to_s)
             )
           when :schema
-            return Main.result_single_object(Transfer::Spec::SCHEMA.merge({'$comment'=>'DO NOT EDIT, this file was generated from the YAML.'}))
+            schema = Transfer::Spec::SCHEMA.merge({'$comment'=>'DO NOT EDIT, this file was generated from the YAML.'})
+            agent = options.get_next_argument('transfer agent name', mandatory: false)
+            schema['properties'] = schema['properties'].select{ |_k, v| CommandLineBuilder.supported_by_agent(agent, v)} unless agent.nil?
+            schema['properties'] = schema['properties'].sort.to_h
+            return Main.result_single_object(schema)
           when :errors
             error_data = []
             Transfer::ERROR_INFO.each_pair do |code, prop|