+++ CHANGELOG.md
@@ -7,7 +7,7 @@
     * `aoc`: #196 Command `packages list` now also supports option `once_only`.
     * `vault`: Support for IBM HashiCorp Vault to store secrets.
     * `wizard`: Preset name can now be specified as optional positional parameter.
-    * `config`: New command `ascp schema` displays JSON schema for transfer spec.
+    * `config`: New command `ascp schema` displays JSON schema for transfer spec for all, or just one agent.
   * Issues Fixed:
     * `aoc`: #195 `package receive ALL` for shared inbox without workspace now works.
   * Breaking Changes: